Simplifies the nature of references.
They require no special syntax (names starting with $) anymore.
Every normal ident (word, consisting of one or more letters) is
identified as a stitch name. If the parser does not recognize
a stitch name as a builtin, it is emitted as a Reference node.

This simplifies the lexer and allows us to treat stitch names
the same way as pattern names. As far as the parser is concerned,
they are no different in their functionality. A stitch represents
one or more sets of stitch primitives. So does a pattern.